+++ spree/api/app/serializers/spree/api/v3/market_serializer.rb
@@ -16,7 +16,7 @@ module Spree
 
         many :countries,
              resource: Spree.api.country_serializer,
-             if: proc { include?(:countries) }
+             if: proc { expand?(:countries) }
       end
     end
   end